src/main/java/dateStructure/dsPlay/dsa/setAndMap/BSTSet.java
package dateStructure.dsPlay.dsa.setAndMap;

import dateStructure.dsPlay.dsa.tree.BST;

public class BSTSet<E extends Comparable<E>> implements Set<E>{

    private BST<E> bst;

    public BSTSet() {
        bst = new BST<>();
    }

    @Override
    public void add(E e) {
        bst.add(e);
    }

    @Override
    public void remove(E e) {
        bst.remove(e);
    }

    @Override
    public boolean contains(E e) {
        return bst.contains(e);
    }

    @Override
    public int geiSize() {
        return bst.getSize();
    }

    @Override
    public boolean isEmpty() {
        return bst.isEmpty();
    }

    @Override
    public String toString() {
        return bst.levelOrder().toString();
    }
}
